+++ pkg/agentcompose/exec_service.go
@@ -2,6 +2,7 @@ package agentcompose
 
 import (
 	execdomain "agent-compose/internal/agentcompose/exec"
+	execservice "agent-compose/internal/agentcompose/exec/service"
 	rundomain "agent-compose/internal/agentcompose/run"
 	appconfig "agent-compose/internal/config"
 	"context"
@@ -45,80 +46,13 @@ func (s *Service) ExecStream(ctx context.Context, req *connect.Request[agentcomp
 type execStreamSender func(*agentcomposev2.ExecStreamResponse) error
 
 func (s *Service) executeProjectCommand(ctx context.Context, req *agentcomposev2.ExecRequest, execID string, send execStreamSender) (*agentcomposev2.ExecResult, error) {
-	if s.store == nil || s.configDB == nil || s.runtimes == nil {
-		return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("exec runtime dependencies are required"))
+	executor := execservice.ProjectCommandExecutor{
+		Config:        s.config,
+		Store:         s.store,
+		Runtimes:      s.runtimes,
+		ResolveTarget: s.resolveExecTargetSession,
 	}
-	session, runID, err := s.resolveExecTargetSession(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	command := strings.TrimSpace(req.GetCommand().GetCommand())
-	if command == "" {
-		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("exec command is required"))
-	}
-	if send != nil {
-		if err := send(&agentcomposev2.ExecStreamResponse{
-			EventType: agentcomposev2.ExecStreamEventType_EXEC_STREAM_EVENT_TYPE_STARTED,
-			ExecId:    execID,
-			SessionId: session.Summary.ID,
-			RunId:     runID,
-		}); err != nil {
-			return nil, connect.NewError(connect.CodeUnknown, err)
-		}
-	}
-	appconfig.ApplyDefaultGuestPaths(s.config)
-	vmState, err := s.store.GetVMState(session.Summary.ID)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, err)
-	}
-	runtime, err := s.runtimes.ForSession(session)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, err)
-	}
-	accumulator := execStreamAccumulator{}
-	var sendErr error
-	writer := func(chunk ExecChunk) {
-		if sendErr != nil {
-			return
-		}
-		accumulator.writeChunk(chunk)
-		if send != nil {
-			sendErr = send(&agentcomposev2.ExecStreamResponse{
-				EventType: agentcomposev2.ExecStreamEventType_EXEC_STREAM_EVENT_TYPE_OUTPUT,
-				ExecId:    execID,
-				SessionId: session.Summary.ID,
-				RunId:     runID,
-				Chunk:     chunk.Text,
-				IsStderr:  chunk.IsStderr,
-			})
-		}
-	}
-	cwd := strings.TrimSpace(req.GetCwd())
-	if cwd == "" {
-		cwd = s.config.GuestWorkspacePath
-	}
-	execCtx, cancel := execContext(ctx, req.GetTimeoutMs())
-	defer cancel()
-	result, execErr := runtime.ExecStream(execCtx, session, vmState, ExecSpec{
-		Command: command,
-		Args:    append([]string(nil), req.GetCommand().GetArgs()...),
-		Env:     execEnvMap(req.GetEnv()),
-		Cwd:     cwd,
-	}, writer)
-	if sendErr != nil {
-		return nil, connect.NewError(connect.CodeUnknown, sendErr)
-	}
-	if execErr != nil {
-		result = mergeExecResults(result, accumulator.result(firstNonZeroInt(result.ExitCode, 1), false))
-		result.ExitCode = firstNonZeroInt(result.ExitCode, 1)
-		result.Success = false
-		if strings.TrimSpace(result.Output) == "" {
-			result.Output = firstNonEmpty(result.Stderr, result.Stdout, execErr.Error())
-		}
-		return execResultResponse(execID, session.Summary.ID, runID, req, cwd, result, execErr), nil
-	}
-	result = mergeExecResults(result, accumulator.result(result.ExitCode, result.Success))
-	return execResultResponse(execID, session.Summary.ID, runID, req, cwd, result, nil), nil
+	return executor.Execute(ctx, req, execID, execservice.StreamSender(send))
 }
 
 func (s *Service) resolveExecTargetSession(ctx context.Context, req *agentcomposev2.ExecRequest) (*Session, string, error) {
